+++ misc/metrics/CHANGELOG.md
@@ -1,4 +1,20 @@
-# 0.10.1 [unreleased]
+# 0.11.0 [unreleased]
+
+- Update to `libp2p-dcutr` `v0.8.0`.
+
+- Update to `libp2p-identify` `v0.41.0`.
+
+- Update to `libp2p-relay` `v0.14.0`.
+
+- Update to `libp2p-core` `v0.38.0`.
+
+- Update to `libp2p-swarm` `v0.41.0`.
+
+- Update to `libp2p-ping` `v0.41.0`.
+
+- Update to `libp2p-kad` `v0.42.0`.
+
+- Update to `libp2p-gossipsub` `v0.43.0`.
 
 - Added `protocol_stack` metrics. See [PR 2982].