+++ transports/dns/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.38.0 [unreleased]
+
+- Update to `libp2p-core` `v0.38.0`.
+
 # 0.37.0
 
 - Remove default features. If you previously depended on `async-std` you need to enable this explicitly now. See [PR 2918].